remove copied-over cosmos provider tests, rename processor cosmos -> penumbra
