+++ relayer/chains/penumbra/penumbra_chain_processor.go
@@ -1,4 +1,4 @@
-package cosmos
+package penumbra
 
 import (
 	"context"
